final housekeeping

+++ c_progs/dice/dice.c
@@ -22,12 +22,10 @@ void generate(uint16_t c, uint16_t ma, uint16_t mi){
 	do{
 		//generate number between min - max
 		num = (rand() % (max - min + 1)) + min;
-		//printf("%d\n",num);
+		
 		//check the number based on index of the check map
 		if(checkmap[(num-min)] == false){
 			memcpy(&arr[inclusions],&num,sizeof(uint16_t));
-			//arr[inclusions] = num;
-			//printf("%d\n",arr[inclusions]);
 			checkmap[(num-min)] = true;
 			inclusions++;
 		}
@@ -35,7 +33,7 @@ void generate(uint16_t c, uint16_t ma, uint16_t mi){
 
 	printf("\n______________________________________\nPrinting spewage:\n\n");
 
-	for(int n = 0; n < count; n++){
+	for(uint16_t n = 0; n < count; n++){
 		printf("%d:%d\t",n+1,arr[n]);
 	}
